프론트를 통해 로그인, 회원가입 테스트(토큰을 저장해서 헤더에 담아 넘겨줘야 다른 요청들을 확인 할 수 있음.)

+++ src/main/java/com/example/springboot_vue_h2db/controller/user/UserController.java
@@ -8,18 +8,10 @@ import com.example.springboot_vue_h2db.model.response.SingleResult;
 import com.example.springboot_vue_h2db.repo.UserRepository;
 import com.example.springboot_vue_h2db.service.ResponseService;
 import io.swagger.annotations.*;
-import io.swagger.models.auth.In;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
 
 @Api(tags = {"2. User"})
 @RequiredArgsConstructor // final로 선언된 객체에 대해서 생성자 주입을 수행한다. 이걸 사용하지 않고 @Autowired를 사용해도 됨.
@@ -43,9 +35,9 @@ public class UserController {
             @ApiImplicitParam(name = "X-AUTH-TOKEN", value = "로그인 성공 후 access_token", required = false, dataType = "String", paramType = "header")
     })
     @ApiOperation(value = "회원 단건 조회", notes = "userId로 회원 조회")
-    @GetMapping(value = "/user/")
+    @GetMapping(value = "/user/") // request가 들어오면 security filter를 먼저 거친다.
     public SingleResult<User> findUserById(@ApiParam(value = "언어", defaultValue = "ko") @RequestParam String lang) {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication(); // Bean(전역에 선언된) SecurityContextHolder에서 현재 인증된 사용자의 정보를 가져온다.
         String id = authentication.getName();
         return responseService.getSingleResult(userRepository.findByUid(id).orElseThrow(CustomNotFoundException::new));
     }
@@ -55,7 +47,7 @@ public class UserController {
     })
     @ApiOperation(value = "회원 수정", notes = "회원 정보 수정")
     @PutMapping(value = "/user")
-    public SingleResult<User> modify(@ApiParam(value = "회원번호", required = true) @RequestParam Integer msrl,
+    public SingleResult<User> modify(@ApiParam(value = "회원번호", required = true) @RequestParam Long msrl,
                                      @ApiParam(value = "회원아이디", required = true) @RequestParam String uid,
                                      @ApiParam(value = "회원이름", required = true) @RequestParam String name) {
         User user = User.builder()
